backend-production/requirements-production.txt
amqp==5.2.0
anyio==4.0.0
APScheduler==3.10.1
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==3.6.4.0
blinker==1.8.1
Brotli==1.1.0
CacheControl==0.13.1
cachetools==5.3.1
celery==5.2.7
certifi==2021.10.8
cffi==1.15.1
channels==4.1.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap4==2023.1
cron_descriptor==1.4.0
cryptography==43.0.0
cssselect2==0.7.0
daphne==4.1.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-channels==0.7.0
django-compat==1.0.15
django-cors-headers==3.14.0
django-crispy-forms==2.1
django-import-export==3.2.0
django-jazzmin==3.0.0
django-mathfilters==1.0.0
django-minio-backend==3.6.0
django-redis==5.3.0
django-soft-delete==0.9.21
django-softdelete==0.10.5
django-timezone-field==6.1.0
django-universal-logger==0.0.3
django-weasyprint==2.2.1
django4-background-tasks==1.2.9
djangorestframework==3.15.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.2.0
firebase-admin==6.2.0
Flask==3.0.3
fonttools==4.44.0
gevent==23.9.1
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
greenlet==3.0.3
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
hyperlink==21.0.0
idna==3.3
importlib_metadata==7.1.0
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.3
jsonfield==3.1.0
jsonrpcserver==5.0.9
jsonschema==4.17.3
kombu==5.3.4
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.5
minio==7.2.8
msgpack==1.0.5
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
OSlash==0.6.3
packaging==24.1
pandas==2.2.0
Pillow==9.5.0
prompt-toolkit==3.0.43
proto-plus==1.22.3
protobuf==4.24.1
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydyf==0.8.0
pyHanko==0.18.0
pyhanko-certvalidator==0.22.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.1
pypdf==3.9.1
pyphen==0.14.0
pypng==0.20220715.0
pyrsistent==0.19.3
pyTelegramBotAPI==4.14.0
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.20.0
python-telegram-bot==20.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
redis==4.5.5
reportlab==3.6.13
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.1.0
six==1.16.0
sentry-sdk[django]
sniffio==1.3.0
sqlparse==0.4.2
svglib==1.5.1
sweetify==2.3.1
tablib==3.5.0
tinycss2==1.2.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2024.1
tzlocal==4.2
ujson==5.5.0
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.9
vine==5.1.0
wcwidth==0.2.12
weasyprint==60.1
webencodings==0.5.1
Werkzeug==3.0.2
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
zipp==3.18.1
zope.event==5.0
zope.interface==6.0
zopfli==0.2.3
pycountry==24.6.1
ip2geotools==0.1.6  